🚚 refactor: shared/view-mode를 entities/member로 흡수
[변경 내용]
- shared/view-mode 11개 파일을 entities/member/view-mode로 이동
- 소비자 17곳(sidebar, project, matching, application, routes)의 import 경로를
  @/shared/view-mode -> @/entities/member/view-mode로 치환

[변경 이유]
- view-mode의 모든 파일이 member identity(MemberInfoResponse, isAnyOperator,
  useMe, ChallengerInfoResponse)에서 파생되는 "뷰어 관점" 도메인 로직이라
  shared에 있으면 shared -> entities 역방향 위반을 12건 유발
- member와 강결합이라 형제 entity로 분리하면 cross-entity 위반이 되므로,
  member 엔티티의 세그먼트로 흡수하는 것이 정확(authStore가 이미 member에
  있는 것과 일관 - member 엔티티가 "뷰어의 정체성과 관점"을 소유)
- boundaries 위반 105 -> 95건 (-10), shared -> entities 17 -> 6건

[고려 사항]
- entities/member -> view-mode 참조는 me.test.ts(테스트 파일)뿐이고 테스트는
  boundaries 검사 제외 대상이라 순환 제약 없음. 이동 후 member 내부 참조가 되어
  same-slice로 합법
- view-mode 내부의 @/entities/member/* 절대 경로는 이동 후 same-slice 참조가
  되어 그대로 유효(상대 경로 변환 불필요)
- 남은 shared 역방향 12건(Profile/ProfileDropdown, api/me, useSchoolChapterMap
  등)은 다음 커밋에서 정리
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
