+++ src/components/sidebar/useVisibleSidebarSections.ts
@@ -5,8 +5,8 @@ import {
   isAnyOperator,
   isCurrentTermPm,
 } from "@/entities/member/model/identity"
+import { useViewerIdentity } from "@/entities/member/view-mode/useViewerIdentity"
 import { SIDEBAR_ITEMS } from "@/shared/config/navigation"
-import { useViewerIdentity } from "@/shared/view-mode/useViewerIdentity"
 
 import { filterSectionsByPermission } from "./utils"
