+++ src/entities/member/view-mode/availableViewModes.test.ts
