+++ src/entities/member/view-mode/availableViewModes.ts
