+++ src/entities/member/view-mode/viewModeParts.ts
