+++ src/routes/matching/projects/announce/index.tsx
@@ -12,6 +12,7 @@ import {
   isChapterPresident,
   isSuperAdmin,
 } from "@/entities/member/model/identity"
+import { useViewerIdentity } from "@/entities/member/view-mode/useViewerIdentity"
 import {
   getAllChapters,
   getAllGisu,
@@ -35,7 +36,6 @@ import { Button } from "@/shared/ui/Button"
 import { MarkdownRenderer } from "@/shared/ui/MarkdownRenderer"
 import { Pagination } from "@/shared/ui/Pagination"
 import { SegmentButton } from "@/shared/ui/segment-button/SegmentButton"
-import { useViewerIdentity } from "@/shared/view-mode/useViewerIdentity"
 
 interface AnnounceSearch {
   chapter: Chapter
